+++ packages/eyes-webdriverio-5/src/fluent/IgnoreRegionBySelector.js
@@ -1,7 +1,12 @@
 'use strict'
 
-const {GetRegion, CoordinatesType, Location, Region} = require('@applitools/eyes-sdk-core')
-
+const {
+  GetRegion,
+  CoordinatesType,
+  Location,
+  Region,
+  locatorToPersistedRegions,
+} = require('@applitools/eyes-sdk-core')
 const {SelectorByLocator} = require('./SelectorByLocator')
 
 class IgnoreRegionBySelector extends GetRegion {
@@ -45,6 +50,10 @@ class IgnoreRegionBySelector extends GetRegion {
    */
   async getSelector(eyes) {
     return new SelectorByLocator(this._selector).getSelector(eyes)
+  }
+
+  async toPersistedRegions(driver) {
+    return locatorToPersistedRegions(this._selector, driver)
   }
 }